@@ -38,7 +38,6 @@
     "eslint-plugin-react": "^7.4.0",
     "fetch-mock": "^5.13.0",
     "handlebars": "^4.0.10",
-    "iltorb": "^1.3.9",
     "istanbul": "^0.4.5",
     "karma": "^1.7.1",
     "karma-chai": "^0.1.0",
@@ -92,15 +91,12 @@
     "postcss-cli": "^4.1.1",
     "postcss-discard-duplicates": "^2.1.0",
     "postcss-extract-css-block": "^0.1.0",
-    "preact": "^8.2.5",
-    "preact-compat": "^3.17.0",
     "promise-rat-race": "^1.5.1",
-    "proptypes": "^1.1.0",
     "raw-loader": "^0.5.1",
     "semver": "^5.4.1",
     "shellpromise": "^1.4.0",
     "watch-run": "^1.2.5",
-    "webpack": "^3.6.0",
+    "webpack": "^3.8.0",
     "webpack-merge": "^4.1.0",
     "webpack-sources": "^1.0.1",
     "webpack-stats-plugin": "^0.1.5"
@@ -134,5 +130,6 @@
       "path": "./public/n-ui/n-ui-core.css",
       "threshold": "8.4 Kb"
     }
-  ]
+  ],
+  "false": {}
 }
