@@ -7,18 +7,21 @@ const appInfo = {
 };
 
 oErrors.init({
-	enabled: window.nextFlags.clientErrorReporting && appInfo.isProduction,
+	enabled: window.FT.flags.clientErrorReporting && appInfo.isProduction,
 	sentryEndpoint: 'https://62a990fd8dce4a27aafb006b58783f66@sentry.io/195030',
 	siteVersion: appInfo.version,
-	logLevel: window.nextFlags.clientDetailedErrorReporting ? 'contextonly' : 'off',
+	logLevel: window.FT.flags.clientDetailedErrorReporting ? 'contextonly' : 'off',
 	tags: {
 		appName: appInfo.name
+	},
+	filterError: function () {
+		return !window.FT.disableOErrors;
 	},
 	errorBuffer: window.errorBuffer || []
 });
 
 // turn on more detailed error reporting of ajax calls
-if (window.nextFlags.clientAjaxErrorReporting) {
+if (window.FT.flags.clientAjaxErrorReporting) {
 	const realFetch = window.fetch;
 	window.fetch = function (url, opts) {
 		return realFetch.call(this, url, opts)
