@@ -4,7 +4,16 @@ const getDomPath = require('../../../ft/utils/getDomPath');
 
 describe('getDomPath', function () {
 
-	const rootEl = document.body;
+	let rootEl;
+
+	before(() => {
+		rootEl = document.createElement('div');
+		document.body.appendChild(rootEl);
+	});
+
+	after(function () {
+		rootEl.parentNode.removeChild(rootEl);
+	});
 
 	beforeEach(function () {
 		rootEl.innerHTML = null;
