@@ -15,17 +15,21 @@ const assetManager = require('./lib/asset-manager');
 module.exports = options => {
 
 	options = Object.assign({}, {
-		withHandlebars: true,
-		withNavigation: true,
-		withAnonMiddleware: true,
+		// hack: shouldn't be able to turn off, but it makes writing tests SOOO much easier
 		withAssets: true,
+		withHandlebars: true,
+
 		withJsonLd: false,
-		withFlags: true,
 		withBackendAuthentication: true,
 		withServiceMetrics: true,
 		product: '',
 		layoutsDir: path.join(__dirname, '../browser/layout'),
-	}, options || {});
+	}, options || {}, {
+		// the options below are forced to be on
+		withNavigation: true,
+		withAnonMiddleware: true,
+		withFlags: true
+	});
 
 	const {app, meta, addInitPromise} = nExpress.getAppContainer(options);
 
@@ -76,7 +80,7 @@ module.exports = options => {
 
 	if (options.withJsonLd) {
 		app.use(function (req, res, next) {
-			if (res.locals.flags && res.locals.flags.newSchema) {
+			if (res.locals.flags.newSchema) {
 				res.locals.jsonLd = [nextJsonLd.webPage()];
 			}
 			next();
