@@ -20,7 +20,7 @@ const anonModels = {
 };
 
 function showFirstClickFree (req, res) {
-	return res.locals.flags && res.locals.flags.firstClickFree &&
+	return res.locals.flags.firstClickFree &&
 			req.get('FT-Access-Decision') === 'GRANTED' &&
 			req.get('FT-Access-Decision-Policy') === 'PRIVILEGED_REFERER_POLICY';
 }
